FixedPoint: partial support for negative values.

+++ test/FixedPoint.h
@@ -6,9 +6,9 @@
 template <int kFraction = 8>
 uint32_t ToFixedPoint(float Value)
 {
-    uint32_t Result = static_cast<uint32_t>(Value * pow(2, kFraction));
+    int32_t Result = static_cast<int32_t>(Value * pow(2, kFraction));
 
-    return Result;
+    return static_cast<uint32_t>(Result);
 }
 
 template <int kFraction = 8>
